requirements.txt
Django>=4.2,<5.0
python-telegram-bot==20.7
telethon==1.33.1
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.0
django-environ==0.11.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
daphne==4.0.0
nest-asyncio==1.6.0